refactor: adjust jwt token provider into its own package
