+++ src/main/java/com/jpcchaves/adotar/application/service/jwt/contracts/JwtTokenProvider.java
@@ -1,0 +1,14 @@
+package com.jpcchaves.adotar.application.service.jwt.contracts;
+
+import org.springframework.security.core.Authentication;
+
+public interface JwtTokenProvider {
+    String generateToken(Authentication authentication);
+
+    String getUsernameFromToken(String token);
+
+    boolean validateToken(String token);
+
+    String getClaimFromTokenByKey(String token,
+                                  String key);
+}
