+++ src/main/java/com/jpcchaves/adotar/presentation/security/JwtAuthenticationFilter.java
@@ -1,11 +1,13 @@
 package com.jpcchaves.adotar.presentation.security;
 
-import com.jpcchaves.adotar.application.service.JwtTokenProvider;
+import com.jpcchaves.adotar.application.service.jwt.contracts.JwtTokenProvider;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+
 import java.io.IOException;
+
 import org.springframework.http.HttpHeaders;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.context.SecurityContextHolder;
@@ -18,55 +20,55 @@ import org.springframework.web.filter.OncePerRequestFilter;
 
 @Component
 public class JwtAuthenticationFilter extends OncePerRequestFilter {
-  private final JwtTokenProvider jwtTokenProvider;
-  private final UserDetailsService userDetailsService;
+    private final JwtTokenProvider jwtTokenProvider;
+    private final UserDetailsService userDetailsService;
 
-  public JwtAuthenticationFilter(
-      JwtTokenProvider jwtTokenProvider,
-      UserDetailsService userDetailsService) {
-    this.jwtTokenProvider = jwtTokenProvider;
-    this.userDetailsService = userDetailsService;
-  }
-
-  @Override
-  protected void doFilterInternal(
-      HttpServletRequest request,
-      HttpServletResponse response,
-      FilterChain filterChain)
-      throws ServletException, IOException {
-    String token = getTokenFromRequest(request);
-
-    if (isTokenValid(token)) {
-      String username = jwtTokenProvider.getUsernameFromToken(token);
-      UserDetails userDetails =
-          this.userDetailsService.loadUserByUsername(username);
-
-      UsernamePasswordAuthenticationToken authenticationToken =
-          new UsernamePasswordAuthenticationToken(
-              userDetails, null, userDetails.getAuthorities());
-
-      authenticationToken.setDetails(
-          new WebAuthenticationDetailsSource().buildDetails(request));
-
-      SecurityContextHolder.getContext().setAuthentication(authenticationToken);
+    public JwtAuthenticationFilter(
+            JwtTokenProvider jwtTokenProvider,
+            UserDetailsService userDetailsService) {
+        this.jwtTokenProvider = jwtTokenProvider;
+        this.userDetailsService = userDetailsService;
     }
 
-    filterChain.doFilter(request, response);
-  }
+    @Override
+    protected void doFilterInternal(
+            HttpServletRequest request,
+            HttpServletResponse response,
+            FilterChain filterChain)
+            throws ServletException, IOException {
+        String token = getTokenFromRequest(request);
 
-  private String getTokenFromRequest(HttpServletRequest request) {
-    String token = request.getHeader(HttpHeaders.AUTHORIZATION);
-    if (checkIfHeaderHasBearerToken(token)) {
-      return token.substring(7);
+        if (isTokenValid(token)) {
+            String username = jwtTokenProvider.getUsernameFromToken(token);
+            UserDetails userDetails =
+                    this.userDetailsService.loadUserByUsername(username);
+
+            UsernamePasswordAuthenticationToken authenticationToken =
+                    new UsernamePasswordAuthenticationToken(
+                            userDetails, null, userDetails.getAuthorities());
+
+            authenticationToken.setDetails(
+                    new WebAuthenticationDetailsSource().buildDetails(request));
+
+            SecurityContextHolder.getContext().setAuthentication(authenticationToken);
+        }
+
+        filterChain.doFilter(request, response);
     }
-    return null;
-  }
 
-  private Boolean checkIfHeaderHasBearerToken(String token) {
-    return StringUtils.hasText(token) && token.startsWith("Bearer ");
-  }
+    private String getTokenFromRequest(HttpServletRequest request) {
+        String token = request.getHeader(HttpHeaders.AUTHORIZATION);
+        if (checkIfHeaderHasBearerToken(token)) {
+            return token.substring(7);
+        }
+        return null;
+    }
 
-  private Boolean isTokenValid(String token) {
-    return StringUtils.hasText(token) && jwtTokenProvider.validateToken(token);
-  }
+    private Boolean checkIfHeaderHasBearerToken(String token) {
+        return StringUtils.hasText(token) && token.startsWith("Bearer ");
+    }
+
+    private Boolean isTokenValid(String token) {
+        return StringUtils.hasText(token) && jwtTokenProvider.validateToken(token);
+    }
 }
